# back-end/src/main/java/com/achref/authpost/post/PostController.java
package com.achref.authpost.post;

import jakarta.validation.*;
import lombok.RequiredArgsConstructor;
import io.swagger.v3.oas.annotations.Parameter;
import org.springframework.security.core.Authentication;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;


@RestController
@RequestMapping("post")
@RequiredArgsConstructor
@Tag(name = "post")
public class PostController {

    private final PostService service;

    @PostMapping("/add-post")
    public ResponseEntity<Long> savePost(
            @Valid @RequestBody PostRequest request,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(service.save(request, connectedUser));
    }

    @GetMapping("/posts")
    public ResponseEntity<?> findAllPosts(
            @RequestParam(name = "page", defaultValue = "0", required = false) int page,
            @RequestParam(name = "size", defaultValue = "5", required = false) int size,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(service.findAllPosts(page, size, connectedUser));
    }
}